+++ packages/Webkul/ImageCache/src/Exceptions/InvalidTemplate.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Webkul\ImageCache\Exceptions;
+
+class InvalidTemplate extends \Exception
+{
+    /**
+     * Create an instance.
+     */
+    public function __construct(string $themeCode, string $name, string $template)
+    {
+        parent::__construct("Image template [{$name}] of theme [{$themeCode}] must be a class with an applyFilter() method, [{$template}] given.");
+    }
+}